Update to postgresql 13 and postgis 3
Npm scripts have been refactored so that the "wait-port" package, which
is a development dependency, is not used when running "npm run migrate"
(which is used in production). The new "npm run migrate:wait" script can
be used in development to both wait for the database and migrate it,
while "npm run migrate" simply triggers the migration without waiting.